refactor(ui): split plugin lifecycle detail renderer
