+++ ui/src/pages/plugins/settings-view.ts
@@ -34,7 +34,6 @@ import {
 import { clawHubPackageUrl } from "./catalog-links.ts";
 import { formatCompactCount } from "./catalog-results.ts";
 import {
-  pluginOriginLabel,
   renderArtTile,
   renderPluginDeclaredCapabilities,
   renderPluginGrants,
@@ -43,6 +42,7 @@ import { renderPluginDetailShell } from "./detail-shell.ts";
 import { buildInstalledPluginDetailTabs, type InstalledPluginDetailTab } from "./detail-tabs.ts";
 import { renderPluginOfficialBadge } from "./plugin-card.ts";
 import { pluginRowKey, type PluginRowMessage } from "./plugin-row-message.ts";
+import { renderPluginLifecycle } from "./settings-lifecycle.ts";
 import { pluginEntryValue } from "./settings-model.ts";
 
 export type PluginSettingsTab = "installed" | "advanced";
@@ -85,7 +85,7 @@ type InventoryProps = SharedProps & {
   onOpenPlugin: (pluginId: string) => void;
 };
 
-type DetailProps = SharedProps & {
+export type DetailProps = SharedProps & {
   pluginId: string;
   inspection: PluginsInspectResult | null;
   inspectionError: string | null;
@@ -434,106 +434,6 @@ function renderAccess(props: DetailProps): TemplateResult {
   `;
 }
 
-function renderLifecycle(props: DetailProps, plugin: PluginCatalogItem): TemplateResult {
-  const key = pluginRowKey(plugin.id);
-  const source = props.inspection?.source;
-  const trust = props.inspection?.trust;
-  const rows = html`
-    ${renderSettingsRow({
-      title: t("pluginsPage.detailPluginId"),
-      control: html`<code>${plugin.id}</code>`,
-      carapace: true,
-    })}
-    ${
-      plugin.version
-        ? renderSettingsRow({
-            title: t("pluginsPage.version"),
-            control: html`<span>${`v${plugin.version}`}</span>`,
-            carapace: true,
-          })
-        : nothing
-    }
-    ${
-      plugin.packageName
-        ? renderSettingsRow({
-            title: t("pluginsPage.detailPackage"),
-            control: html`<code>${plugin.packageName}</code>`,
-            carapace: true,
-          })
-        : nothing
-    }
-    ${
-      plugin.origin
-        ? renderSettingsRow({
-            title: t("pluginsPage.detailOrigin"),
-            control: html`<span>${pluginOriginLabel(plugin.origin)}</span>`,
-            carapace: true,
-          })
-        : nothing
-    }
-    ${
-      source
-        ? renderSettingsRow({
-            title: t("pluginsPage.installedSource"),
-            control: html`<span>${source.spec ?? source.packageName ?? source.kind}</span>`,
-            carapace: true,
-          })
-        : nothing
-    }
-    ${
-      source?.integrity
-        ? renderSettingsRow({
-            title: t("pluginsPage.integrity"),
-            control: html`<code title=${source.integrity}>${source.integrity.slice(0, 20)}…</code>`,
-            carapace: true,
-          })
-        : nothing
-    }
-    ${
-      trust
-        ? renderSettingsRow({
-            title: t("pluginsPage.trustStatus"),
-            control: html`<span>${trust.disposition}</span>`,
-            carapace: true,
-          })
-        : nothing
-    }
-    ${
-      plugin.removable
-        ? renderSettingsRow({
-            title: t("pluginsPage.uninstall"),
-            description: t("pluginsPage.uninstallDescription"),
-            control: renderReasonedDisabledControl(
-              props.mutationBlockedReason,
-              html`<button
-                type="button"
-                class="btn danger oc-action oc-action-secondary"
-                ?disabled=${
-                  !props.mutationBlockedReason && (!props.canMutate || Boolean(props.busy[key]))
-                }
-                aria-disabled=${!props.canMutate ? "true" : nothing}
-                aria-label=${t("pluginsPage.uninstallNamed", { name: plugin.name })}
-                @click=${() => {
-                  if (props.canMutate && !props.busy[key]) {
-                    props.onUninstall(plugin.id, key);
-                  }
-                }}
-              >
-                ${t("pluginsPage.uninstall")}
-              </button>`,
-            ),
-            carapace: true,
-          })
-        : renderSettingsRow({
-            title: t("pluginsPage.uninstall"),
-            description: t("pluginsPage.managedCannotUninstall"),
-            carapace: true,
-          })
-    }
-  `;
-  return rows;
-}
-
 function renderInstalledComponentRows(values: readonly string[]): TemplateResult {
   return html`<div class="plugin-catalog-detail__rows">
     ${values.map(
@@ -628,7 +528,7 @@ function renderInstalledTabPanel(
     return renderAccess(props);
   }
   if (tab === "lifecycle") {
-    return renderLifecycle(props, plugin);
+    return renderPluginLifecycle(props, plugin);
   }
   return renderInstalledAdvanced(props);
 }
